requirements.txt
Flask
deep-translator
python-docx
PyMuPDF
